refactor: replace uint with Dimension and update related logic for improved type safety

+++ day_12/internal/abstractions/presents.go
@@ -2,7 +2,7 @@ package abstractions
 
 const (
 	// MaximumShapeSize /* All the presents occupies a 3x3 region */
-	MaximumShapeSize = 3
+	MaximumShapeSize = uint(3)
 )
 
 type Presents struct {
